Move tenant aware entity in the general domain package
