+++ project/src/main/java/caselab/domain/entity/Document.java
@@ -2,7 +2,6 @@ package caselab.domain.entity;
 
 import caselab.domain.entity.enums.DocumentStatus;
 import caselab.elastic.listener.DocumentEventListener;
-import caselab.multitenancy.domain.entity.TenantAwareEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EntityListeners;
