+++ project/src/main/java/caselab/domain/entity/DocumentEvent.java
@@ -1,7 +1,6 @@
 package caselab.domain.entity;
 
 import caselab.domain.entity.enums.EventType;
-import caselab.multitenancy.domain.entity.TenantAwareEntity;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
