+++ project/src/main/java/caselab/domain/entity/DocumentVersion.java
@@ -1,7 +1,6 @@
 package caselab.domain.entity;
 
 import caselab.domain.entity.attribute.value.AttributeValue;
-import caselab.multitenancy.domain.entity.TenantAwareEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
