+++ project/src/main/java/caselab/domain/entity/RefreshToken.java
@@ -1,6 +1,5 @@
 package caselab.domain.entity;
 
-import caselab.multitenancy.domain.entity.TenantAwareEntity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
