+++ project/src/main/java/caselab/domain/entity/Subscription.java
@@ -1,6 +1,5 @@
 package caselab.domain.entity;
 
-import caselab.multitenancy.domain.entity.TenantAwareEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
