+++ project/src/main/java/caselab/domain/entity/document/type/to/attribute/DocumentTypeToAttribute.java
@@ -2,7 +2,7 @@ package caselab.domain.entity.document.type.to.attribute;
 
 import caselab.domain.entity.Attribute;
 import caselab.domain.entity.DocumentType;
-import caselab.multitenancy.domain.entity.TenantAwareEntity;
+import caselab.domain.entity.TenantAwareEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.EmbeddedId;
 import jakarta.persistence.Entity;
